Make library more reflection-friendly
I find it easier to work with reflect.Method than hardcoding method signatures. At least it is on the code I'm working on :)

I could, of course, extract the signature from the reflection, but it turned out to be surprinsingly cumbersome.

So instead, I reworked YAHFA to use the reflected method. Of course.
I also added some glue (`findMethod`)  to keep it compatible with whatever is already using this library

And, while I was at it, I also added some checks and exception throwing.
(At first it took me surprinsigly long time to figure out that I had wrong arguments on my hook)