Fix `checkCompatibleMethods()` check on Backup
When replacing A with B, the arguments of A must be assignable from the arguments of B, and the result of B must be assignable with the result of B.

This check is correctly implemented for replacing `target` with `hook`, but is reversed when replacing `backup` with `target`.

It seems is counter-intuitive that we would have different arguments on the target, hook and backup methods. Most of the time we don't, but it can be useful for corner cases, while respecting strong typing.